menambahkan register, login, getuserbytokencode

+++ internal/model/model.go
@@ -1,7 +1,7 @@
 package model
 
 type ApiResponse[T any] struct {
-	Code int8 `json:"code"`
+	Code int32 `json:"code"`
 	Status string `json:"status"`
 	Data T `json:"data"`
 }
